Khadija

+++ app/Http/Controllers/Api/V1/Dashboard/RasidJobController.php
@@ -6,7 +6,6 @@ use Illuminate\Http\Request;
 use App\Models\RasidJob\RasidJob;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Dashboard\RasidJob\{RasidJobResource, RasidJobCollection};
-use App\Http\Resources\Dashboard\ActivityLogResource;
 use App\Http\Requests\V1\Dashboard\RasidJobRequest;
 use App\Http\Requests\V1\Dashboard\ReasonRequest;
 
@@ -96,10 +95,23 @@ class RasidJobController extends Controller
     }
 
 
-    public function destroy(ReasonRequest $request, RasidJob $rasidJob)
+    public function destroy( RasidJob $rasidJob)
     {
 
+        if(!$rasidJob->is_vacant){
+
+            return response()->json([
+                'status' => false,
+                'message' => trans("dashboard.rasid_job.jobs_hired_archived"),
+                'data' => null
+            ], 422);
+
+
+        }
+
         $rasidJob->delete();
+
+
 
         return RasidJobResource::make($rasidJob)
             ->additional([
@@ -108,10 +120,22 @@ class RasidJobController extends Controller
             ]);
     }
 
-    public function forceDelete(ReasonRequest $request, $id)
+    public function forceDelete($id)
     {
 
         $rasidJob = RasidJob::onlyTrashed()->findOrFail($id);
+
+        if(!$rasidJob->is_vacant){
+
+            return response()->json([
+                'status' => false,
+                'message' => trans("dashboard.rasid_job.jobs_hired_deleted"),
+                'data' => null
+            ], 422);
+
+
+        }
+
         $rasidJob->forceDelete();
 
         return RasidJobResource::make($rasidJob)
